Started with my battle to reduce the almighty god of god classes to a shivering pulp. - Abstract GameStateHelper class to help create a set of helpers so I can reduce the stress on GameState. - GameStatePlayerHelper to handle all player methods - GameStateCirclesHelper to hande all circle/gate methods for now - GameStateInterfaceHelper to handle all UI/menu/user-interface methods
<organisation,layout
