+++ src/main/java/gui/GameStateInterfaceHelper.java
@@ -1,0 +1,41 @@
+package gui;
+
+import org.newdawn.slick.GameContainer;
+import org.newdawn.slick.Graphics;
+
+/**
+ * GameState Helper class for managing the UI. This is done to prevent
+ * GameState from having too much responsibility and/or knowledge. The class
+ * should only be used in conjunction with GameState.
+ * @author Mark
+ */
+public class GameStateInterfaceHelper extends GameStateHelper {
+
+	/**
+	 * Constructor method for creating GameStateInterfaceHelper object.
+	 */
+	public GameStateInterfaceHelper() {
+		
+	}
+	
+	@Override
+	public void enter() {
+		
+	}
+
+	@Override
+	public void exit() {
+		
+	}
+
+	@Override
+	public void update(GameContainer container, float deltaFloat) {
+		
+	}
+
+	@Override
+	public void render(Graphics graphics, GameContainer container) {
+		
+	}
+
+}
