+++ src/main/java/logic/CircleList.java
@@ -1,6 +1,13 @@
 package logic;
 
+import gui.RND;
+
 import java.util.ArrayList;
+
+import org.newdawn.slick.Color;
+import org.newdawn.slick.Graphics;
+import org.newdawn.slick.Image;
+import org.newdawn.slick.SlickException;
 
 /**
  * Class containing a list of bouncingcircles.
@@ -12,6 +19,20 @@ public class CircleList {
 	private ArrayList<BouncingCircle> circles;
 	private int highestID;
 	
+	private Image[] ballsImagesN;
+	private Image[] ballsImagesA;
+
+	private static final int BALL_IMAGE_THREE = 3;
+	private static final int BALL_IMAGE_FOUR = 4;
+	private static final int BALL_IMAGE_FIVE = 5;
+	private static final int AMOUNT_OF_BALLS = 6;
+	private static final int CIRCLE_DRAW_OFFSET = 13;
+	private static final int MINIMUM_RADIUS = 10;
+	private static final int RADIUS_2 = 20;
+	private static final int RADIUS_3 = 30;
+	private static final int RADIUS_4 = 45;
+	private static final int RADIUS_5 = 65;
+	private static final int RADIUS_6 = 90;
 	
 	
 	/**
@@ -25,6 +46,25 @@ public class CircleList {
 			if (circle.getId() > highestID) {
 				highestID = circle.getId();
 			}
+		}
+		
+		ballsImagesN = new Image[AMOUNT_OF_BALLS];
+		try {
+			ballsImagesN[0] = new Image("resources/images_Balls/Ball_90_Norm.png");
+			ballsImagesN[1] = new Image("resources/images_Balls/Ball_65_Norm.png");
+			ballsImagesN[2] = new Image("resources/images_Balls/Ball_45_Norm.png");
+			ballsImagesN[BALL_IMAGE_THREE] = new Image("resources/images_Balls/Ball_30_Norm.png");
+			ballsImagesN[BALL_IMAGE_FOUR] = new Image("resources/images_Balls/Ball_20_Norm.png");
+			ballsImagesN[BALL_IMAGE_FIVE] = new Image("resources/images_Balls/Ball_10_Norm.png");
+			ballsImagesA = new Image[AMOUNT_OF_BALLS];
+			ballsImagesA[0] = new Image("resources/images_Balls/Ball_90_Add.png");
+			ballsImagesA[1] = new Image("resources/images_Balls/Ball_65_Add.png");
+			ballsImagesA[2] = new Image("resources/images_Balls/Ball_45_Add.png");
+			ballsImagesA[BALL_IMAGE_THREE] = new Image("resources/images_Balls/Ball_30_Add.png");
+			ballsImagesA[BALL_IMAGE_FOUR] = new Image("resources/images_Balls/Ball_20_Add.png");
+			ballsImagesA[BALL_IMAGE_FIVE] = new Image("resources/images_Balls/Ball_10_Add.png");
+		} catch (SlickException e) {
+			e.printStackTrace();
 		}
 		
 	}
@@ -86,14 +126,12 @@ public class CircleList {
 		return highestID;
 	}
 	
-	
 	/**
 	 * @return the circles
 	 */
 	public ArrayList<BouncingCircle> getCircles() {
 		return circles;
 	}
-	
 	
 	/**
 	 * @param circles the circles to set
@@ -102,5 +140,39 @@ public class CircleList {
 		this.circles = circles;
 	}
 	
+	/**
+	 * Render all the circles in the circlelist to the screen.
+	 * @param graphics context to draw in.
+	 * @param color to draw circles with.
+	 */
+	public void drawCircles(Graphics graphics, Color color) {
+		for (BouncingCircle circle : circles) { 
+			int r = (int) circle.getRadius(), offset = CIRCLE_DRAW_OFFSET;
+			final float xPosition = circle.getMinX() - offset;
+			final float yPosition = circle.getMinY() - offset;
+			switch (r) {
+			case(RADIUS_6) : RND.drawColor(graphics, ballsImagesN[0], ballsImagesA[0],
+					xPosition, yPosition, color); break;
+			case(RADIUS_5) : RND.drawColor(graphics, ballsImagesN[1], ballsImagesA[1],
+					xPosition, yPosition, color); break;
+			case(RADIUS_4) : RND.drawColor(graphics, ballsImagesN[2], ballsImagesA[2],
+					xPosition, yPosition, color); break;
+			case(RADIUS_3) : RND.drawColor(graphics, 
+					ballsImagesN[BALL_IMAGE_THREE], ballsImagesA[BALL_IMAGE_THREE],
+					xPosition, yPosition, color); break;
+			case(RADIUS_2) : RND.drawColor(graphics, 
+					ballsImagesN[BALL_IMAGE_FOUR], ballsImagesA[BALL_IMAGE_FOUR],
+					xPosition, yPosition, color); break;
+			case(MINIMUM_RADIUS) : RND.drawColor(graphics, 
+					ballsImagesN[BALL_IMAGE_FIVE], ballsImagesA[BALL_IMAGE_FIVE],
+					xPosition, yPosition, color); break;
+			default:
+				try {
+					throw new SlickException("Radius was not one of the supported");
+				} catch (SlickException e) {
+					e.printStackTrace(); }
+		}
+		}
+	}
 
 }
